requirements.txt
Django==2.2.2
djangorestframework==3.9.4
openapi-codec==1.3.2
Pillow==6.0.0
pipdeptree==0.13.2
psycopg2==2.8.2
psycopg2-binary==2.8.2
python-dotenv==0.10.3
PyYAML==5.1.2
setuptools==41.0.1
simplejson==3.16.0
wheel==0.33.4